feat: replace npm compiler module with remote one
I'm so so so happy with this migration 😭. It's not much, but I tried adding this feature almost a year ago when I built this playground, but couldn't do it because my programming skills were kinda not here yet, but mostly I didn't write a detailed design of this migration (I wouldn't do that at that time, I'd just think about somthing and write code right away 💀). Using the detailed design document, it was like building brick wall - every step fell into place effortlessly.

Just a note for myself.